qspi flash 10M speed

+++ CM7/Core/Inc/inFlash.h
@@ -8,7 +8,7 @@
 #ifndef SRC_INFLASH_H_
 #define SRC_INFLASH_H_
 
-uint32_t Internal_ReadFlash(uint32_t addrStart, uint32_t *pData, uint32_t dataLen);
-uint32_t Internal_WriteFlash(uint32_t addrStart, const uint32_t *pData, uint32_t dataLen);;
+uint32_t Internal_ReadFlash(uint32_t addrStart, uint8_t *pData, uint32_t dataLen);
+uint32_t Internal_WriteFlash(uint32_t addrStart, const uint8_t *pData, uint32_t dataLen);;
 
 #endif /* SRC_INFLASH_H_ */
